// app/api/models/post.model.ts
import { PostImage } from "./post-image.model";

export interface Post {
    kind: string;
    data: {
      approved_at_utc?: any;
      subreddit: string;
      selftext: string;
      user_reports: any[];
      saved: boolean;
      mod_reason_title?: any;
      gilded: number;
      clicked: boolean;
      title: string;
      link_flair_richtext: any[];
      subreddit_name_prefixed: string;
      hidden: boolean;
      pwls: number;
      link_flair_css_class?: any;
      downs: number;
      thumbnail_height: number;
      parent_whitelist_status: string;
      hide_score: boolean;
      name: string;
      quarantine: boolean;
      link_flair_text_color: string;
      author_flair_background_color?: any;
      subreddit_type: string;
      ups: number;
      domain: string;
      media_embed: {};
      thumbnail_width: number;
      author_flair_template_id?: any;
      is_original_content: boolean;
      secure_media?: any;
      is_reddit_media_domain: boolean;
      category?: any;
      secure_media_embed: {};
      link_flair_text?: any;
      can_mod_post: boolean;
      score: number;
      approved_by?: any;
      thumbnail: string;
      edited: boolean;
      author_flair_css_class?: any;
      author_flair_richtext: any[];
      post_hint: string;
      content_categories?: any;
      is_self: boolean;
      mod_note?: any;
      created: number;
      link_flair_type: string;
      wls: number;
      post_categories?: any;
      banned_by?: any;
      author_flair_type: string;
      contest_mode: boolean;
      selftext_html?: any;
      likes?: any;
      suggested_sort?: any;
      banned_at_utc?: any;
      view_count?: any;
      archived: boolean;
      no_follow: boolean;
      is_crosspostable: boolean;
      pinned: boolean;
      over_18: boolean;
      preview: {
        images: [
          {
              source: PostImage;
            resolutions: PostImage[];
            variants: {};
            id: string;
          }
        ];
        enabled: boolean;
      };
      media_only: boolean;
      can_gild: boolean;
      spoiler: boolean;
      locked: boolean;
      author_flair_text?: any;
      rte_mode: string;
      visited: boolean;
      num_reports?: any;
      distinguished?: any;
      subreddit_id: string;
      mod_reason_by?: any;
      removal_reason?: any;
      id: string;
      report_reasons?: any;
      author: string;
      num_crossposts: number;
      num_comments: number;
      send_replies: boolean;
      mod_reports: any[];
      author_flair_text_color?: any;
      permalink: string;
      whitelist_status: string;
      stickied: boolean;
      url: string;
      subreddit_subscribers: number;
      created_utc: number;
      media?: any;
      is_video: boolean;
    };
  }